libcxx: allow tuning of __deque_block_size
Allocating 4096 for a std::deque with only 1 to maybe 16 entries is a
poor use of memory.  Allocating chunk dropped to 384 for embedded
builds.  This saved 10K overall in the DMT-TRIPLE project.

So we introduce _DEQUESIZE to allow external tuning of the std::deque
block size.  Setting the default to 384 for embedded systems, and
leave it alone at page size 4096 for __linux__ systems.

NB: on windows version of the deque header, they offer _DEQUESIZ
    (notice no trailing E) for a similar purpose.

Signed-off-by: Mark Salyzyn <mark.salyzyn@gmail.com>